Make InputGetter act as remembered
I used to think I should let disabling this component hang the inputs--but I remembered it as being the one to _clear_ them. So because my project code doesn't rely on this behavior (yet), I reversed it. I believe this is how it should be--the player's code should do what _they_ want when it's disabled.

Changed this as I was writing the input-getter quick-start guide.